params weren't passed from clusterise to tokenise
Line 205/206: instead of passing variadic arguments, they were converted to a list and assigned to the second argument ("strict"). Fixed that now. `[:-1]` is needed because locals() lists also the function groupit() in the end, which we don't want to pass on to tokenise. Renamed "merge()" to "groupit()", because "merge" is already the name of an argument. Didn't want to name it "group()" since it occurs elsewhere in the script already.